use render throttling
