+++ packages/state/package.json
@@ -52,6 +52,9 @@
 			"node_modules/(?!(nanoid)/)"
 		]
 	},
+	"dependencies": {
+		"@tldraw/utils": "workspace:*"
+	},
 	"devDependencies": {
 		"@types/lodash": "^4.14.188",
 		"@types/react": "^18.2.47",
