+++ packages/state/src/lib/react/index.ts
@@ -1,6 +1,7 @@
 export { track } from './track'
 export { useAtom } from './useAtom'
 export { useComputed } from './useComputed'
+export { useLayoutReaction } from './useLayoutReaction'
 export { useQuickReactor } from './useQuickReactor'
 export { useReactor } from './useReactor'
 export { useStateTracking } from './useStateTracking'
