+++ packages/state/src/lib/react/useLayoutReaction.ts
@@ -1,0 +1,11 @@
+import React from 'react'
+import { useTrackedScheduler } from './useTrackedScheduler'
+
+/** @internal */
+export function useLayoutReaction<T>(name: string, effect: () => void): void {
+	const scheduler = useTrackedScheduler(name, effect)
+
+	React.useLayoutEffect(() => {
+		scheduler.execute()
+	})
+}
