+++ packages/state/src/lib/react/useValue.ts
@@ -1,4 +1,5 @@
 /* eslint-disable prefer-rest-params */
+import { throttleToNextFrame } from '@tldraw/utils'
 import { useMemo, useRef, useSyncExternalStore } from 'react'
 import { Signal, computed, react } from '../core'
 
@@ -81,10 +82,16 @@ export function useValue() {
 		const { subscribe, getSnapshot } = useMemo(() => {
 			return {
 				subscribe: (listen: () => void) => {
-					return react(`useValue(${name})`, () => {
-						$val.get()
-						listen()
-					})
+					return react(
+						`useValue(${name})`,
+						() => {
+							$val.get()
+							listen()
+						},
+						{
+							scheduleEffect: throttleToNextFrame,
+						}
+					)
 				},
 				getSnapshot: () => $val.get(),
 			}
